Dependencies fix
Assuming that .Net Framework 4.5 is to work with Mono I keep
System.Xml.Linq 3.5.* because it seems to be built for that, but it will
force developpers to install .Net Framework 3.5 to make it works. So I
add the System.Xml.Linq 4.0.* to avoid this situation for Microsoft
Visual Studio 2015.